chore: move get nododb to other file
